+++ examples/reference/state-abi-bin-to-json.ts
@@ -15,7 +15,7 @@ async function main() {
 
     console.log("State ABI bin -> json response", prettifyJson(response))
   } catch (error) {
-    console.log("An error occurred", prettifyJson(error))
+    console.log("An error occurred", error)
   }
 }