+++ examples/reference/state-key-accounts.ts
@@ -11,7 +11,7 @@ async function main() {
 
     console.log("State key accounts response", prettifyJson(response))
   } catch (error) {
-    console.log("An error occurred", prettifyJson(error))
+    console.log("An error occurred", error)
   }
 }