+++ src/types/http-client.ts
@@ -1,5 +1,7 @@
 export const V1_AUTH_ISSUE = "/v1/auth/issue"
 
+export const V0_FETCH_TRANSACTION = "/v0/transactions/:id"
+
 export const V0_STATE_ABI = "/v0/state/abi"
 export const V0_STATE_ABI_BIN_TO_JSON = "/v0/state/abi/bin_to_json"
 
@@ -9,7 +11,6 @@ export const V0_STATE_PERMISSION_LINKS = "/v0/state/permission_links"
 export const V0_STATE_TABLE = "/v0/state/table"
 export const V0_STATE_TABLES_SCOPES = "/v0/state/tables/scopes"
 export const V0_STATE_TABLES_ACCOUNTS = "/v0/state/tables/accounts"
-
 export const V0_STATE_TABLE_SCOPES = "/v0/state/table_scopes"
 
 export const V0_SEARCH_TRANSACTIONS = "/v0/search/transactions"